requirements.txt
google-cloud-storage==2.11.0
google-cloud-bigquery==3.10.0